refactor(package) 패키지 구성 변경
- https://github.com/next-step/ddd-strategic-design/pull/320 리뷰를 토대로 패키지 재구성
- 누락된 port/out 추가
