+++ src/main/java/kitchenpos/eatinorder/adapter/in/EatinTableRestController.java
@@ -1,0 +1,9 @@
+package kitchenpos.eatinorder.adapter.in;
+
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RequestMapping("/api/eatin-tables")
+@RestController
+public class EatinTableRestController {
+}
