+++ src/main/java/kitchenpos/eatinorder/domain/EatinOrder.java
@@ -1,0 +1,4 @@
+package kitchenpos.eatinorder.domain;
+
+public class EatinOrder {
+}
