--- src/main/java/kitchenpos/eatinorder/eatinorder/adapter/in/EatinOrderRestController.java
@@ -1,9 +1,0 @@
-package kitchenpos.eatinorder.eatinorder.adapter.in;
-
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RequestMapping("/api/eatin-orders")
-@RestController
-public class EatinOrderRestController {
-}
