--- src/main/java/kitchenpos/eatinorder/eatinorder/application/port/in/EatinOrderUseCase.java
@@ -1,5 +1,0 @@
-package kitchenpos.eatinorder.eatinorder.application.port.in;
-
-public interface EatinOrderUseCase {
-
-}
