--- src/main/java/kitchenpos/eatinorder/eatinorder/domain/EatinOrder.java
@@ -1,4 +1,0 @@
-package kitchenpos.eatinorder.eatinorder.domain;
-
-public class EatinOrder {
-}
