--- src/main/java/kitchenpos/eatinorder/eatinorder/domain/OrderLineItem.java
@@ -1,4 +1,0 @@
-package kitchenpos.eatinorder.eatinorder.domain;
-
-public class OrderLineItem {
-}
