--- src/main/java/kitchenpos/eatinorder/eatintable/domain/EatInTable.java
@@ -1,5 +1,0 @@
-package kitchenpos.eatinorder.eatintable.domain;
-
-public class EatInTable {
-
-}
