+++ src/main/java/kitchenpos/menu/adapter/in/MenuRestController.java
@@ -1,0 +1,10 @@
+package kitchenpos.menu.adapter.in;
+
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RequestMapping("/api/menus")
+@RestController
+public class MenuRestController {
+
+}
