+++ src/main/java/kitchenpos/menu/domain/Menu.java
@@ -1,0 +1,5 @@
+package kitchenpos.menu.domain;
+
+public class Menu {
+
+}
