+++ src/main/java/kitchenpos/menu/domain/MenuGroup.java
@@ -1,0 +1,5 @@
+package kitchenpos.menu.domain;
+
+public class MenuGroup {
+
+}
