+++ src/main/java/kitchenpos/menu/domain/MenuProduct.java
@@ -1,0 +1,5 @@
+package kitchenpos.menu.domain;
+
+public class MenuProduct {
+
+}
