--- src/main/java/kitchenpos/menu/menu/adapter/in/MenuRestController.java
@@ -1,10 +1,0 @@
-package kitchenpos.menu.menu.adapter.in;
-
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RequestMapping("/api/menus")
-@RestController
-public class MenuRestController {
-
-}
