--- src/main/java/kitchenpos/menu/menu/application/port/in/MenuUseCase.java
@@ -1,5 +1,0 @@
-package kitchenpos.menu.menu.application.port.in;
-
-public interface MenuUseCase {
-
-}
