--- src/main/java/kitchenpos/menu/menu/domain/Menu.java
@@ -1,5 +1,0 @@
-package kitchenpos.menu.menu.domain;
-
-public class Menu {
-
-}
