--- src/main/java/kitchenpos/menu/menu/domain/MenuProduct.java
@@ -1,5 +1,0 @@
-package kitchenpos.menu.menu.domain;
-
-public class MenuProduct {
-
-}
