--- src/main/java/kitchenpos/menu/menugroup/application/MenuGroupService.java
@@ -1,7 +1,0 @@
-package kitchenpos.menu.menugroup.application;
-
-import kitchenpos.menu.menugroup.application.port.in.MenuGroupUseCase;
-
-public class MenuGroupService implements MenuGroupUseCase {
-
-}
