--- src/main/java/kitchenpos/menu/menugroup/application/port/in/MenuGroupUseCase.java
@@ -1,5 +1,0 @@
-package kitchenpos.menu.menugroup.application.port.in;
-
-public interface MenuGroupUseCase {
-
-}
