--- src/main/java/kitchenpos/menu/menugroup/domain/MenuGroup.java
@@ -1,5 +1,0 @@
-package kitchenpos.menu.menugroup.domain;
-
-public class MenuGroup {
-
-}
