--- src/main/java/kitchenpos/menu/product/adapter/in/ProductRestController.java
@@ -1,10 +1,0 @@
-package kitchenpos.menu.product.adapter.in;
-
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RequestMapping("/api/products")
-@RestController
-public class ProductRestController {
-
-}
