--- src/main/java/kitchenpos/menu/product/application/port/in/ProductUseCase.java
@@ -1,5 +1,0 @@
-package kitchenpos.menu.product.application.port.in;
-
-public interface ProductUseCase {
-
-}
