+++ src/main/java/kitchenpos/product/adapter/in/ProductRestController.java
@@ -1,0 +1,10 @@
+package kitchenpos.product.adapter.in;
+
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RequestMapping("/api/products")
+@RestController
+public class ProductRestController {
+
+}
