+++ src/main/java/kitchenpos/product/application/port/in/ProductUseCase.java
@@ -1,0 +1,5 @@
+package kitchenpos.product.application.port.in;
+
+public interface ProductUseCase {
+
+}
