+++ src/main/java/kitchenpos/product/domain/Product.java
@@ -1,0 +1,5 @@
+package kitchenpos.product.domain;
+
+public class Product {
+
+}
